requirements.txt
faiss-cpu
fastapi
openai
python-dotenv
requests
streamlit
uvicorn